refactor(envoy-filter): move everything into same package
