+++ privacy-profile-composer/cmd/envoy-filter/main.go
@@ -3,13 +3,13 @@ package main
 import (
 	"github.com/envoyproxy/envoy/contrib/golang/filters/http/source/go/pkg/http"
 
-	"privacy-profile-composer/pkg/envoy_filter/config"
+	"privacy-profile-composer/pkg/envoyfilter"
 )
 
 const Name = "simple"
 
 func init() {
-	http.RegisterHttpFilterConfigFactoryAndParser(Name, config.ConfigFactory, &config.Parser{})
+	http.RegisterHttpFilterConfigFactoryAndParser(Name, envoyfilter.ConfigFactory, &envoyfilter.ConfigParser{})
 }
 
 func main() {}
